[kexts] Move AMD Kexts to Extra and disable them
AMDRyzenCPUPowerManagement is supposed to handle the entire power management
of the AMD CPU, though even with the Kext not being loaded, though - most
likely due to SSDT-PLUG - IORegistryExplorer reports a properly populated
C000 CPU node with X86PlatformPlugin and CPU P-States et al. While testing,
it seems the processor and the overall system runs cooler as the system fan
even gets disabled after cooling down all components (which wasn't the case
with AMDRyzenCPUPowerManagement). Performance under load remains about the
same. So, as there seem to be reports that AMDRyzenCPUPowerManagement might
render the system unstable, just disable it for now.

As SMCAMDProcessor depends on AMDRyzenCPUPowerManagement, disable that one
as well.